128-bit checksum, block type in block header

+++ src/segment/block/checksum.rs
@@ -2,12 +2,12 @@
 // This source code is licensed under both the Apache 2.0 and MIT License
 // (found in the LICENSE-* files in the repository)
 
-/// An 64-bit checksum
+/// An 128-bit checksum
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct Checksum(u64);
+pub struct Checksum(u128);
 
 impl std::ops::Deref for Checksum {
-    type Target = u64;
+    type Target = u128;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -16,7 +16,7 @@ impl std::ops::Deref for Checksum {
 
 impl Checksum {
     #[must_use]
-    pub fn from_raw(value: u64) -> Self {
+    pub fn from_raw(value: u128) -> Self {
         Self(value)
     }
 }
